Now HuntingStrategy only returns the desired position, all the rest of the hunt logic is in the hunt state, only works with GWOStrategy for the moment
