+++ EcoSim/Assets/Model/Hunting/HuntingStrategy.cs
@@ -2,5 +2,6 @@
 
 public abstract class HuntingStrategy
 {
-    public abstract void Hunt(Animal agent, Dictionary<int, Animal> friendly, Dictionary<int, Animal> foes);
+    public int FramesUpdate;
+    public abstract Vec3 GetDesiredPosition(Animal agent, List<Vec3> friendlyPositions, Animal prey);
 }
