+++ EcoSim/Assets/SimulationData/SimulationDataPhase2/Stage2-Simulation1.txt
@@ -32,3 +32,7 @@ Iteracion|InicialPresas|InicialPredadores|SupervivientesPresas|SupervivientesPre
 0|50|6|50|6|1000|950
 0|50|6|48|6|1000|952
 1|72|9|72|9|1142.4|1070.4
+0|50|6|50|6|1000|950
+0|50|6|50|6|1000|950
+0|50|6|49|6|1000|951
+1|73|6|71|6|1141.2|1070.2
