+++ EcoSim/Assets/SimulationData/SimulationDataPhase2/Stage2-Simulation2.txt
@@ -5,3 +5,4 @@ Iteracion|InicialPresas|InicialPredadores|SupervivientesPresas|SupervivientesPre
 0|50|5|47|5|1000|953
 0|50|6|50|6|1000|950
 0|50|6|48|6|1000|952
+0|50|6|50|6|1000|950
